==> data/Sources/compilation_mining_data-master/compilation_mining_data-master/02_scripts/01_detailed_data/02_conversion.Rmd <==
---
title: "harmonization"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r setup , include = FALSE, echo = FALSE, message = FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(rlang)
library(DT)

# clear R environment
rm(list = ls())

```



```{r , echo = FALSE, message = FALSE}
## read files
# harmonized data file
detailed <- read_rds("./03_intermediate/01_detailed_data/all_tables_harmonized.rds")

# conversion factors
conv_fac <- read_delim("./01_input/04_factors/conversion_factors.csv", delim = ";")

# list of columns
sheets_columns <- read_delim("./01_input/02_lists_and_concordance_tables/01_detailed_data/sheets_columns.csv", delim = ";")







#### conversion of units --------------

# select all relevant columns for conversion and replacement
sel_col <- sheets_columns %>% 
  filter(!is.na(conv_unit_col)) %>%
  filter(sheet != "water_input")



# create dummy tables for no_match entries
no_match_conv_fac <- tibble(unit_id = NA, material_id = NA, waste_type = NA) %>% 
  mutate(across(where(is.logical), as.character)) %>%
  filter(!is.na(unit_id))



# for checks:
# i <- "general"
# j <- "surface_area"
# k <- "unit_surface_area"



## convert all relevant columns in a loop
  ## (including two nested loops, one for value columns based on grades, second for grade columns)
for (i in sel_col %>% distinct(sheet) %>% pull()) {
  
  sheet <- detailed[[i]]
  
  # filter for only the relevant columns in this sheet
  sel_col_filtered <- sel_col %>% filter(sheet == i)
  
  
  
  # first nested loop for each variable in the current sheet, first checking, then converting it based on unit
    # converting and replacing units is done in two nested loops because some columns share the same unit column, e.g. in sheet "commodities" and "waste"
    # (other possibility would have been to add a unit column for each, but this option has been discarded for now)
  
  for (j in sel_col_filtered$variable) {
  
  unit_col <- sel_col_filtered %>% filter(variable == j) %>% pull(conv_unit_col)
  mat_col <- sel_col_filtered %>% filter(variable == j) %>% pull(conv_mat_col)
  
  
  
  ## check for units and material-unit/waste_type-unit combinations existing in current sheet, but not in conv_fac
  # could be checked at the beginning, but more efficient here, as info on unit_cat is already included at this point
  # IDs could also simply be checked against unit_id table, but that would be lacking a check for unit-material combinations

  # check for units of mass, grade, area
  a1 <- setdiff(
    sheet %>% 
      distinct(!!parse_expr(unit_col)) %>% 
      rename(unit_id = !!parse_expr(unit_col)),
    conv_fac %>% 
      distinct(unit_id)
    ) %>% 
    add_column(material_id = NA, waste_type = NA) %>% mutate(across(where(is.logical), as.character)) %>%
    filter(!is.na(unit_id))
  
  # check for material-unit combinations of volume and energy (differentiated for waste sheet and all other sheets)
  if(!is.na(mat_col)
     ) if(i == "waste") {
          a2 <- setdiff(
            sheet %>%
              distinct(!!parse_expr(unit_col), waste_type) %>% 
              rename(unit_id = !!parse_expr(unit_col)) %>%
              filter(!unit_id %in% (conv_fac %>% filter(unit_cat %in% c("mass", "grade", "area")) %>% select(unit_id) %>% pull())),
            conv_fac %>% 
              filter(unit_cat %in% c("volume", "energy")) %>%
              distinct(unit_id, waste_type)
            ) %>% 
            add_column(material_id = NA) %>% mutate(across(where(is.logical), as.character)) %>%
            filter(!is.na(unit_id))
          } else a2 <- setdiff(
              sheet %>% 
                distinct(!!parse_expr(unit_col), !!parse_expr(mat_col)) %>% 
                rename(unit_id = !!parse_expr(unit_col), material_id = !!parse_expr(mat_col)) %>%
              filter(!unit_id %in% (conv_fac %>% filter(unit_cat %in% c("mass", "grade", "area")) %>% select(unit_id) %>% pull())),
              conv_fac %>% 
                filter(unit_cat %in% c("volume", "energy")) %>%
                distinct(unit_id, material_id)
              ) %>% 
              add_column(waste_type = NA) %>% mutate(across(where(is.logical), as.character)) %>%
              filter(!is.na(unit_id))
  
  
  # combine individual no_match tables and add to dummy table
    # (double entries possible if more than one variable share same unit column)
  no_match_conv_fac <- no_match_conv_fac %>%
    union(., a1) %>%
    { if(!is.na(mat_col)
         ) union(., a2) else .
      } %>%
    distinct()
  
  
  
  # add respective factors to table, first for mass, grade, area, then for energy and volume
    # including adding all rows where unit was NA (which are thrown out during filtering for unit categories)
  sheet_p1 <- sheet %>%
    left_join(.,
              conv_fac %>%
                distinct(unit_cat, unit_id),
              by = setNames(nm = unit_col, "unit_id")
    ) %>%
    filter(unit_cat %in% c("mass", "grade", "area")) %>%
    select(-unit_cat) %>%
    left_join(.,
              conv_fac %>%
                select(-material_id, -waste_type),
              by = setNames(nm = unit_col, "unit_id")
              ) %>%
    union(.,
      sheet %>%
    filter(is.na(!!parse_expr(unit_col))) %>%
    add_column(unit_cat = NA) %>% mutate(across(where(is.logical), as.character)) %>%
    add_column(conv_factor_t = NA, conv_factor_ppm = NA, conv_factor_sqkm = NA) %>% mutate(across(where(is.logical), as.double))
    )
  
  
  # define check whether volume and energy units are included (and therefore material_id or waste_type are relevant as well)
    # which is used several times below
  exist_vol_ene <- sheet %>%
      distinct(!!parse_expr(unit_col)) %>% filter(!is.na(.)) %>% pull() %in% (conv_fac %>% filter(unit_cat %in% c("volume", "energy")) %>% distinct(unit_id) %>% pull())
  
  if(TRUE %in% exist_vol_ene
      ) sheet <- sheet_p1 %>%
          union(.,
                sheet %>%
                  left_join(.,
                            conv_fac %>%
                              distinct(unit_cat, unit_id),
                            by = setNames(nm = unit_col, "unit_id")
                  ) %>%
                  filter(unit_cat %in% c("volume", "energy")) %>%
                  select(-unit_cat) %>%
                  { if(i == "waste" & j != "total_material")
                      left_join(.,
                                conv_fac %>%
                                  select(-material_id),
                                by = setNames(nm = c(unit_col, mat_col), c("unit_id", "waste_type"))
                                )
                    else if(i == "waste" & j == "total_material")
                      left_join(.,
                                conv_fac %>%
                                  filter(is.na(material_id)) %>% # this works because there is only one entry with NA for each unit
                                  select(-material_id, -waste_type),
                                by = setNames(nm = unit_col, "unit_id")
                      )
                    else 
                      left_join(.,
                                conv_fac %>%
                                  select(-waste_type),
                                by = setNames(nm = c(unit_col, mat_col), c("unit_id", "material_id"))
                      )
                      }
          ) else sheet <- sheet_p1
  
  
  # convert j (i.e. respective unit column) based on unit_cat and conversion factors joined above
  sheet <- sheet %>%
    mutate(
      !!j := case_when(
        unit_cat %in% c("mass", "volume", "energy") ~ !!parse_expr(j) * conv_factor_t,
        unit_cat == "grade" ~ !!parse_expr(j) * conv_factor_ppm,
        unit_cat == "area" ~ !!parse_expr(j) * conv_factor_sqkm
      )
    )
  
  
  
  # deselect joined columns to enable join without problem in next loop
  sheet <- sheet %>% select(-unit_cat, -conv_factor_t, -conv_factor_ppm, -conv_factor_sqkm)
  
  }
  
  
  
  
  
  # second nested loop for each unit variable in the current sheet, replacing unit to either "t" or "ppm"
  for (k in sel_col_filtered %>% distinct(conv_unit_col) %>% pull()) {
    
    sheet <- sheet %>%
      left_join(.,
                conv_fac %>% distinct(unit_cat, unit_id), 
                by=setNames(nm = k, "unit_id")
      ) %>%
      mutate(
        !!k := case_when(
          unit_cat %in% c("mass", "volume", "energy") ~ "t",
          unit_cat == "grade" ~ "ppm",
          unit_cat == "area" ~ "sq.km"
        )
      ) %>% 
      select(-unit_cat)
    
  }
  
  
  
  
  detailed[[i]] <- sheet
  
}


#create directory
dir.create("./04_output/01_detailed_data/02_conversion/01_no_match_tables/", showWarnings = FALSE)

# save no_match table as csv
write_delim(no_match_conv_fac, "./04_output/01_detailed_data/02_conversion/01_no_match_tables/no_match_conv_fac.csv", delim = ";")

# table for Rmd output
datatable(no_match_conv_fac, rownames = FALSE, caption = "Missing conversion factors")

```



```{r , echo = FALSE, message = FALSE}

#### conversion by shares, based on production or reserves share -----------------

# select all relevant columns for conversion
sel_col <- sheets_columns %>% 
  filter(!is.na(conv_by_share)) %>%
  filter(sheet != "water_input")



# convert all relevant columns in loops again
for (i in sel_col %>% distinct(sheet) %>% pull()) {
  
  sheet <- detailed[[i]]
  
  # filter for only the relevant columns in this sheet
  sel_col_filtered <- sel_col %>% filter(sheet == i)
  
  # loop for each variable in sheet
  for (j in sel_col_filtered$variable) {
    
    share_col <- sel_col_filtered %>% filter(variable == j) %>% pull(conv_by_share)
    
    # convert values
    sheet <- sheet %>%
      mutate(!!share_col := ifelse(is.na(eval(parse(text = share_col))), 1, (eval(parse(text = share_col))))) %>%
      mutate(!!j := eval(parse(text = j)) / eval(parse(text = share_col)))
    
  }
  
  # change share column to 1 after converting
  sheet <- sheet %>%
    mutate(!!share_col := ifelse(!is.na(eval(parse(text = share_col))), 1, NA))
  
  
  
  detailed[[i]] <- sheet 
  
}



#### save converted tables
write_rds(detailed, "./03_intermediate/01_detailed_data/all_tables_converted.rds")

```

<br>
