==> data/Sources/compilation_mining_data-master/compilation_mining_data-master/02_scripts/01_detailed_data/06_compile_estimation_factors.Rmd <==
---
title: "Compile estimation factors"
output:
  html_document: default
  pdf_document: default
editor_options: 
  chunk_output_type: console
---

```{r setup , include = FALSE, echo = FALSE, message = FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(plotly)
library(scales)
library(grid)
library(knitr)
library(kableExtra)
library(viridis)

# clear R environment
rm(list = ls())

```



```{r , echo = FALSE, message = FALSE}
## read files
# harmonized data file
detailed <- read_rds("./03_intermediate/01_detailed_data/gaps_filled.rds")

# list of columns
sheets_columns <- read_delim("./01_input/02_lists_and_concordance_tables/01_detailed_data/sheets_columns.csv", delim = ";")

# metal prices
met_price <- read_delim("./01_input/05_other_data/average_prices_1990-2020.csv", delim = ";")

# sheets
sheet_com <- detailed$commodities
sheet_min <- detailed$minerals_ores_conce

```

<br> 

```{r , echo = FALSE, message = FALSE}
## include other variables

# countries
sheet_all <- sheet_com %>%
  left_join(.,
            detailed$general %>%
              distinct(mine_fac, alphanumiso) %>%
              mutate(alphanumiso = ifelse(is.na(alphanumiso), "Other", alphanumiso))
  )

# prices
sheet_all <- sheet_all %>%
  left_join(.,
            met_price %>% select(material_id, average_price),
            by = c("commodity" = "material_id")
  )

# ores
sheet_all <- sheet_all %>%
  left_join(.,
            sheet_min %>% 
              filter(type_mineral %in% c("Ore processed")) %>%
              select(mine_fac, min_ore_con, year, value, mine_processing),
            by = c("mine_fac", "min_ore_con", "year", "mine_processing"),
            suffix = c(".com", ".min")
  )



## Calculate stuff

# Calculate value of metal
sheet_est <- sheet_all %>%
  mutate(met_value = value.com * 1000 * average_price) # taking into account proper units: i.e. t vs USD/kg

# Calculating total value per ore, join that into table
sheet_est <- sheet_est %>%
  left_join(
    sheet_est %>%
      group_by(mine_fac, min_ore_con, year, mine_processing) %>%
      summarise(total_met_value = sum(met_value))
  )

# Calculate ratio (t/t) of ore per metal (i.e. share of metal value per ore value times amount of ore, and then "ore share" per amount of metal)
sheet_est <- sheet_est %>%
  mutate(ore_share = (met_value / total_met_value * value.min)) %>%
  mutate(ratio_ore_per_met = ore_share / value.com)



# Remove everything without a ratio (e.g. commodities like alloys or just years lacking data)
  # + everything where value.com and value.min are NA
est_fac_1 <- sheet_est %>%
  filter(
    !is.na(ratio_ore_per_met),
    !is.na(value.com),
    !is.na(value.min)
    ) %>%
  select(alphanumiso, commodity, year, value.com, ratio_ore_per_met)


```

<br> 

```{r , echo = FALSE, message = FALSE, warning = FALSE}

## compile additional factors based on ores vs concentrates
  ## i.e. those cases where only concentrate content has been reported and therefore has not been matched with respective ore in table join
  ## here using amount of ore vs metal content in concentrate

# list of mines with concentrates
mine_list_1 <- sheet_com %>%
  filter(grepl("Con.", min_ore_con)) %>%
  distinct(mine_fac, year)

# list of mines with other commodities than concentrates
mine_list_2 <- sheet_com %>%
  filter(!(grepl("Con.", min_ore_con))) %>%
  distinct(mine_fac, year)

mine_list_con_only <- setdiff(mine_list_1, mine_list_2)


## include commodities, countries, prices, and ores (with four left_joins) + aggregate sheet_com and sheet_min respectively
  # similar to above
  # needs to be done again, because here min_ore_con and mine_processing need to be aggregated in both sheets (as not clear relation between ore and content is given, and therefore double counting would be created during table join)

sheet_all_2 <- mine_list_con_only %>%
  left_join(.,
            sheet_com %>%
              select(mine_fac, commodity, year, value) %>%
              group_by(mine_fac, commodity, year) %>%
              summarise(value = if(all(is.na(value))) NA else sum(value, na.rm = TRUE)) %>%
              ungroup()
            ) %>%
  left_join(.,
            detailed$general %>%
              distinct(mine_fac, alphanumiso) %>%
              mutate(alphanumiso = ifelse(is.na(alphanumiso), "Other", alphanumiso))
            ) %>%
  left_join(.,
            met_price %>% select(material_id, average_price),
            by = c("commodity" = "material_id")
            ) %>%
  left_join(.,
            sheet_min %>% 
              filter(type_mineral %in% c("Ore processed")) %>%
              select(mine_fac, year, value) %>%
              group_by(mine_fac, year) %>%
              summarise(value = if(all(is.na(value))) NA else sum(value, na.rm = TRUE)),
            by = c("mine_fac", "year"),
            suffix = c(".com", ".min")
            )


## Calculate stuff again

# Calculate value of metal
sheet_est_2 <- sheet_all_2 %>%
  mutate(met_value = value.com * 1000 * average_price) # taking into account proper units: i.e. t vs USD/kg

# Calculating total value per ore, join that into table
sheet_est_2 <- sheet_est_2 %>%
  left_join(
    sheet_est_2 %>%
      group_by(mine_fac, year) %>%
      summarise(total_met_value = sum(met_value))
  )

# Calculate ratio (t/t) of ore per metal (i.e. share of metal value per ore value times amount of ore, and then "ore share" per amount of metal)
sheet_est_2 <- sheet_est_2 %>%
  mutate(ore_share = (met_value / total_met_value * value.min)) %>%
  mutate(ratio_ore_per_met = ore_share / value.com)



# Remove everything without a ratio (e.g. commodities like alloys or just years lacking data)
  # + everything where value.com and value.min are NA
est_fac_2 <- sheet_est_2 %>%
  filter(
    !is.na(ratio_ore_per_met),
    !is.na(value.com),
    !is.na(value.min)
    ) %>%
  select(alphanumiso, commodity, year, value.com, ratio_ore_per_met)


```

<br> 

```{r , echo = FALSE, message = FALSE, warning = FALSE}

# Combine both tables and consolidate by country, i.e. group by country and summarize ore ratio with weighted average (weighted by amount of metal)
est_fac_all <- est_fac_1 %>%
  union(.,
        est_fac_2
        ) %>%
  group_by(alphanumiso, commodity, year) %>%
  summarise(ratio_ore_per_met = weighted.mean(ratio_ore_per_met, value.com, na.rm = TRUE)) %>%
  ungroup()



# save data as csv
write_delim(est_fac_all, "./04_output/01_detailed_data/06_est_fac/estimation_factors.csv",
            delim = ";", na = "")

```

<br> 

#### Line charts with time series for each commodity

```{r , echo = FALSE, message = FALSE, warning = FALSE}

# Line charts for checking time series

list_comm <- est_fac_all %>% filter(!is.na(commodity)) %>% distinct(commodity) %>% arrange(commodity) %>% pull()

a <- list()
for (i in list_comm) {

  nr_colour <- nrow(est_fac_all %>% filter(commodity == i) %>% distinct(alphanumiso))

  a[[i]] <- est_fac_all %>%
    filter(commodity == i) %>%
    plotly::plot_ly(
      x = ~year,
      y = ~ratio_ore_per_met,
      color = ~alphanumiso
      ) %>% 
    plotly::add_lines() %>%
    plotly::layout(title = i)
  
}  
  
  htmltools::tagList(a)



```

<br> 

#### All ratios for each commodity and country compared to overall median/mean of commodity

```{r , echo = FALSE, message = FALSE}

# Create dynamic DT::datatable to show highest and lowest values for estimation factors of each commodity (for error checking)
  # And include difference from median/mean of commodity across all years

  highest_values <- est_fac_all %>%
    left_join(
      est_fac_all %>%
        group_by(commodity) %>%
        summarise(
          median_ratio = median(ratio_ore_per_met, na.rm = TRUE),
          mean_ratio = mean(ratio_ore_per_met, na.rm = TRUE)
          ) %>%
        ungroup(),
      by = c("commodity")
      ) %>%
    mutate(
      diff_median_ratio = ratio_ore_per_met / median_ratio,
      diff_mean_ratio = ratio_ore_per_met / mean_ratio
      ) %>%
    arrange(desc(diff_median_ratio))


  # table for Rmd output
DT::datatable(highest_values, rownames = FALSE, caption = "Ratios for each commodity/country/year")  

```
